discord/pyproject.toml
[project]
name = "discord"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "maro114510", email = "atumaro114510@icloud.com" }
]
dependencies = [
    "discord.py>=2.3.2",
    "python-dotenv>=1.0.0",
]
readme = "README.md"
requires-python = ">= 3.8"

[project.scripts]
hello = "discord:hello"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/discord"]
